+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 # Changelog
 
 ## 0.4.3
+- `Macro Explorer` view:
+  - Library nodes use `~/` on descriptions for simplicity in non-Windows platforms.
+  - Fix: Drag-and-drop broken on 0.4.2 with introduction of the `Temporary` node.
 - Update `Git` template.
 - Fix: 'Download .d.ts file' is not a QuickFix.
-- Fix: Drag-and-drop broken on 0.4.2 with introduction of the `Temporary` node.
 
 ## 0.4.2
 - `Macro Explorer` view: